src/test/java/org/simpleframework/croe/BeanContainerTest.java
package org.simpleframework.croe;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class BeanContainerTest {

    private static BeanContainer beanContainer;

    /**
     * 执行测试方法前初始化一次
     */
    @BeforeAll
    static void init(){
        beanContainer = BeanContainer.getInstance();
    }

    @DisplayName("加载目标类及其实例到BeanContainer：loadBeansTest")
    @Test
    public void loadBeansTest(){
        Assertions.assertEquals(false,beanContainer.isLoaded());
        beanContainer.loadBeans("com.lin");
        Assertions.assertEquals(6,beanContainer.size());
        Assertions.assertEquals(true,beanContainer.isLoaded());
    }

    @Test
    void isLoaded() {
    }

    @Test
    void getInstance() {
    }

    @Test
    void loadBeans() {
    }
}
